# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Documentation for API endpoints",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Register new user",
                "operationId": "3b201ddfd948f9d4114befe853f5872a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "name",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "Nugrah"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Fabianugerah Bainasshiddiq"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "rahasia123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success register user"
                    }
                }
            }
        },
        "/api/users/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Login user",
                "operationId": "c9b20e2431404b7383ce7e626b4b6f05",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "Nugrah"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "rahasia123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success login"
                    }
                }
            }
        },
        "/api/users/current": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get current user",
                "operationId": "284ea63f67607acf4715f44787499112",
                "responses": {
                    "200": {
                        "description": "Success get current user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Update current user",
                "operationId": "6b38acce2e0e0be120a452b76d9377b9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fabianugerah"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "rahasia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success update user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/logout": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Logout current user",
                "operationId": "197ceb8d01abe337fbae5165ecca9cd1",
                "responses": {
                    "200": {
                        "description": "Success logout user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}